BUG: Fix RigidAlignmentModuleMetadata
The metadata was no longer being installed to the correct path during
build. I believe this was because of the relative path that moved up
a directory.

The RigidAlignmentModuleMetadata name was changed to have a single
leading underscore to indicate that it is a private folder, per
guidelines in PEP 8.