BUG: Allow running on Windows
The isoformat of datetime has colons between hour, minute, second, so it
is not able to be used as a filename on Windows, although it is fine for
Linux. Replacing : with - allows cross platform filenames.

Also checking the result of the RigidAlignment CLI to prevent a possible
segfault if there is a failure.